responding to PR feedback, rename functions/files
